requirements.txt
requests
matplotlib
numpy
opencv-python
Pillow
mahotas
scikit-image
scipy